wms接口

+++ lyz-app-pojo/src/main/java/cn/com/leyizhuang/app/foundation/pojo/remote/webservice/wms/WtaWarehouseWholeSplitToUnit.java
@@ -41,11 +41,11 @@ public class WtaWarehouseWholeSplitToUnit {
     /**
      * 数量
      */
-    private Double qty;
+    private Integer qty;
     /**
      * 零数量
      */
-    private Double dQty;
+    private Integer dQty;
     /**
      * 状态
      */
